Fixed some stuff up based on warnings from Visual C++: - Channel_MODULE.cc had a line that accidentally referred to pattern_delay_by_frames instead of pattern_delay_frames. G++ didn't bat an eyelid at `+=`-ing a `bool`! - Handled EOF during decode_and_fill in xm_pattern_note in Load_XM.cc. - Moved types containing complex objects out of the #pragma pack(1) declaration in Load_XM.cc. Visual C++ quite rightly pointed out that they could end up with `this` pointers that aren't aligned. - Removed a bunch of copy/paste cruft from dump_pattern in Load_XM.cc. Eliminated all warnings.

+++ formatting.cc
@@ -98,20 +98,20 @@ namespace MultiPLAY
 			d << "name: " << mod->name << endl;
 
 			d << "pattern: " << mod->patterns.size() << endl;
-			for (unsigned i = 0, l = mod->patterns.size(); i < l; i++)
+			for (size_t i = 0, l = mod->patterns.size(); i < l; i++)
 			{
 				d << "[" << i << "]:" << endl;
 				format_pattern(d, &mod->patterns[i]);
 			}
 
 			d << "pattern list: " << mod->pattern_list_length << endl;
-			for (unsigned i = 0, l  = unsigned(mod->pattern_list_length); i < l; i++)
+			for (size_t i = 0, l  = unsigned(mod->pattern_list_length); i < l; i++)
 			{
 				d << "[" << i << "]: ";
 
 				bool found = false;
 
-				for (unsigned j = 0, k = mod->patterns.size(); j < k; j++)
+				for (size_t j = 0, k = mod->patterns.size(); j < k; j++)
 				{
 					if (mod->pattern_list[i] == &mod->patterns[j])
 					{
@@ -126,14 +126,14 @@ namespace MultiPLAY
 			}
 
 			d << "samples: " << mod->samples.size() << endl;
-			for (unsigned i = 0, l = mod->samples.size(); i < l; i++)
+			for (size_t i = 0, l = mod->samples.size(); i < l; i++)
 			{
 				d << "[" << i << "]: ";
 				format_sample(d, mod->samples[i]);
 			}
 
 			d << "sample_info: " << mod->sample_info.size() << endl;
-			for (unsigned i = 0, l = mod->sample_info.size(); i < l; i++)
+			for (size_t i = 0, l = mod->sample_info.size(); i < l; i++)
 				d << "[" << i << "]: vibrato_speed == " << mod->sample_info[i].vibrato_speed << ", vibrato_depth == " << mod->sample_info[i].vibrato_depth << endl;
 
 			d << "channel_enabled: ";
